test(runtime): isolate normalized diagnostics regressions
